+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   allows the single file to include all relevant data.
   CNJR-1649
 - Added Conjur health and info checks for the specified container. CNJR-2397
+- Added container logs to the output archive file, when a container ID is
+  specified. CNJR-2395
 
 ## [0.3.0] - 2023-01-26